change type from int64 to int16

+++ src/face_detection/custom_dataset.py
@@ -20,8 +20,8 @@ class FacesDataset(Dataset):
         for rects in self.frames[dict_key]:
             boxes.append([rects.x, rects.y, rects.x + rects.w, rects.y + rects.h])
 
-        boxes = torch.as_tensor(boxes, dtype=torch.int64)
-        labels = torch.as_tensor(labels, dtype=torch.int64)
+        boxes = torch.as_tensor(boxes, dtype=torch.int16)
+        labels = torch.as_tensor(labels, dtype=torch.int16)
 
         target = {"boxes": boxes, "labels": labels}
